+++ database/migrations/0001_01_01_000000_create_users_table.php
@@ -18,7 +18,7 @@ return new class extends Migration
             $table->timestamp('email_verified_at')->nullable();
             $table->string('password');
             $table->rememberToken();
-            $table->decimal('balance', 10, 2)->default(0.00);
+            $table->bigInteger('balance')->default(0);
             $table->timestamps();
         });
